Address review: move gdp_col regression tests to their own module
